requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2>=2.9,<3.0
django-filter>=23.0,<24.0